Move filter adapters into the ui package
