+++ app/src/main/java/com/detroitlabs/devicemanager/db/DeviceDao.java
@@ -8,6 +8,8 @@ import android.arch.persistence.room.OnConflictStrategy;
 import android.arch.persistence.room.Query;
 import android.arch.persistence.room.Update;
 
+import com.detroitlabs.devicemanager.constants.Platform;
+
 import java.util.List;
 
 @Dao
@@ -32,4 +34,7 @@ public interface DeviceDao {
 
     @Query("SELECT * FROM device WHERE serialNumber = :serialNumber")
     LiveData<Device> getDevice(String serialNumber);
+
+    @Query("SELECT * FROM device WHERE platform in (:platform) AND brandAndModel in (:brandAndModel) AND version in (:version) AND screenSize in (:screenSize) AND screenResolution in (:screenResolution) AND yearClass in (:yearClass) AND isSamsung in (:isSamsung) AND serialNumber <> :thisSerialNumber")
+    LiveData<List<Device>> filterDevice(List<Platform> platform, List<String> brandAndModel, List<String> version, List<String> screenSize, List<String> screenResolution, List<String> yearClass, List<String> isSamsung, String thisSerialNumber);
 }
