+++ app/src/main/java/com/detroitlabs/devicemanager/ui/filter/FilterViewModel.java
@@ -1,0 +1,7 @@
+package com.detroitlabs.devicemanager.ui.filter;
+
+import android.arch.lifecycle.ViewModel;
+
+public class FilterViewModel extends ViewModel {
+
+}
