+++ app/src/main/java/com/detroitlabs/devicemanager/ui/filter/adapters/YearClassAdapter.java
@@ -1,17 +1,17 @@
-package com.detroitlabs.devicemanager.filter.adapters;
-
+package com.detroitlabs.devicemanager.ui.filter.adapters;
 
 import com.detroitlabs.devicemanager.R;
 import com.detroitlabs.devicemanager.constants.FilterType;
 
-public class ScreenSizeAdapter extends FilterOptionAdapter {
+
+public class YearClassAdapter extends FilterOptionAdapter {
     @Override
     public FilterType getFilterType() {
-        return FilterType.SCREEN_SIZE;
+        return FilterType.YEAR_CLASS;
     }
 
     @Override
     public int getTitleRes() {
-        return R.string.screen_size;
+        return R.string.year_class;
     }
 }
